pyproject.toml
[project]
name = "personal-ai-trainer"
version = "0.1.0"
description = "AI-powered personal fitness trainer with Google Drive storage"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "google-generativeai>=0.8.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.110.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    "requests>=2.31.0",
    "extra-streamlit-components>=0.1.71",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
